# AREG-89 Improve performances of the import