The current CSV import of antibodies is slow, there is a couple of fixes made here:

1. the save is not making so much tests at the moment of saving
2. there is a threadpool for trying to save the instances in a parallel way, however the `@atomic` on the `save(...)` method of Antibody probably prevent the threadpool to work properly. I let it anyway, just in case. 

Here is the results I measured for 500 antibodies:

* 500 updates
  * current version: between 18 and 20s
  * this PR: between 10 and 12s
* 500 creates
  * current version: between 15 and 18s
  * this PR: between 9 and 11s

It's faster, but not drastically faster. 

Note 1: When using the threadpool, we cannot remove the `@atomic` on the `save(...)`
Note 2: Not using the threadpool and removing the `@atomic` implies also a little bit of perf improvement, but I'm not sure why the `@atomic` is here in the first place, and if we can safely remove it.